Refactor MSNRegressor and GNNRegressor to subclass OrdinationKNeighborsRegressor
The new OrdinationKNeighborsRegressor base class implements the
KNeighborsRegressor __init__ method with the addition of an n_components
parameter to allow dimensionality reduction via a ComponentReducerMixin
transformer.
